+++ logging.c
@@ -134,7 +134,7 @@ int log_add_plain(FILE *logf, mtype lm, __attribute__((unused)) prio lq, char lp
 int log_add_symbolic(FILE *logf, mtype lm, prio lq, char lp, bool ls, const char *lt, const char *ltag, time_t ts)
 {
 	if(!logf) return(1);
-	fprintf(logf, "u+%lld %s %s %c %c ", (signed long long)ts, mtype_name(lm), prio_name(lq), lp?lp:'0', ls?'>':'<');
+	fprintf(logf, "u+%jd %s %s %c %c ", (intmax_t)ts, mtype_name(lm), prio_name(lq), lp?lp:'0', ls?'>':'<');
 	safeprint(logf, ltag, true);
 	fputc(' ', logf);
 	safeprint(logf, lt, false);
